requirements.txt
asgiref==3.4.1
async-timeout==4.0.3
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.1.0
backports.zoneinfo==0.2.1
certifi==2019.11.28
cffi==1.15.0
channels==3.0.4
channels-redis==4.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
dbus-python==1.2.16
decorator==4.4.2
distro==1.4.0
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.15.2
msgpack==1.1.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.4.0
prompt-toolkit==2.0.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.26.1
pycparser==2.20
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==2.9.0
pyOpenSSL==21.0.0
pytz==2021.3
redis==5.0.8
requests==2.22.0
service-identity==21.1.0
six==1.14.0
sqlparse==0.4.2
ssh-import-id==5.10
traitlets==4.3.3
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.25.8
uWSGI==2.0.20
wcwidth==0.1.8
zope.interface==5.4.0
